Update _config.yml to point to license, not 404
Was previously pointing to https://github.com/hpc-social.github.io/blob/main/LICENSE, which 404's. I assume this was due to the repo previously not being part of an org? 

Regardless, this should have it now point to the correct place in the repos license from within the org. 

Alternative is to point to an embedded GPLv3 license page within the site (already have one here: https://hpc-social.github.io/LICENSE)