updating mastodon links and page
The changes here will link the icon to mastodon directly,
and all links (except for email) now have a _blank target.
The portal and home page link to our Mastodon page where
there is a large obvious button to click to join, so the
user flow should be like Mastodon -> Join Us On Mastodon ->
Mastodon page. I am adding the icon as an svg because I would
have to re-generate the icomoon font and I would rather do
that only if absolutely necessary! This seems to look OK,
granted that I sized it down to match the others.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>